Revert to original test set-up
Move back most of the unit test to a separate file since Travis execution takes to long and exceeds the time limit.

Issues:
#61

Contributors:
@krasen86
