+++ app/src/androidTest/java/se/healthrover/CarManagementTest.java
@@ -24,7 +24,6 @@ import se.healthrover.ui_activity_controller.car_selection.CarSelect;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotEquals;
 
 @RunWith(AndroidJUnit4.class)
 public class CarManagementTest {
@@ -34,6 +33,7 @@ public class CarManagementTest {
     private List<Car> carListTestCars;
     private SqlHelper sqlHelper;
     private CarManagementImp management;
+
 
 
     @Rule
@@ -59,24 +59,6 @@ public class CarManagementTest {
         }
 
     }
-
-    @Test
-    public void updateCarNameTest(){
-        Car car = carListTestCars.get(0);
-        management.updateCarName(car,"test new name", activity);
-        car.setName("test new name");
-        assertEquals(car, sqlHelper.getCarByName("test new name"));
-    }
-
-    @Test
-    public void loadCarsIntoListTest(){
-        management.addCar(carListTestCars.get(0));
-        management.loadCarsIntoList(activity);
-
-        assertEquals(carListTestCars.get(0), management.getCars().get(0));
-    }
-
-
     @Test
     public void getCarNamesTest(){
         String[] result = management.getCarNames();
@@ -94,48 +76,20 @@ public class CarManagementTest {
         assertEquals(car, result);
 
     }
-
     @Test
-    public void getCarByNameTest(){
+    public void updateCarNameTest(){
         Car car = carListTestCars.get(0);
-        String carName = car.getName();
-        Car result = management.getCarByName(carName);
-        assertEquals(carName, result.getName());
-        assertEquals(car, result);
-    }
-    @Test
-    public void getCarsTest(){
-        List<Car> cars = management.getCars();
-        assertEquals(carListTestCars, cars);
+        management.updateCarName(car,"test new name", activity);
+        car.setName("test new name");
+        assertEquals(car, sqlHelper.getCarByName("test new name"));
     }
 
     @Test
-    public void addCarTest(){
-        Car car = new Car("http://" + new Faker().internet().url(), new Faker().name().username());
-        carListTestCars.add(car);
-        int startSize = management.getCars().size();
-        management.addCar(car);
-        assertEquals(startSize + 1 , management.getCars().size());
-        assertEquals(carListTestCars, management.getCars());
-    }
+    public void loadCarsIntoListTest(){
+        management.addCar(carListTestCars.get(0));
+        management.loadCarsIntoList(activity);
 
-    @Test
-    public void removeCarTest(){
-        Car car = management.getCars().get(0);
-        carListTestCars.remove(car);
-        int startSize = management.getCars().size();
-        management.removeCar(car);
-        assertEquals(startSize - 1 , management.getCars().size());
-        assertEquals(carListTestCars, management.getCars());
-    }
-    @Test
-    public void setCarListTest(){
-        for (int i = 1; i < management.getCars().size(); i++){
-            management.getCars().remove(i);
-        }
-        assertNotEquals(carListTestCars, management.getCars());
-        management.setCars(carListTestCars);
-        assertEquals(carListTestCars, management.getCars());
+        assertEquals(carListTestCars.get(0), management.getCars().get(0));
     }
 
     @After
@@ -150,6 +104,5 @@ public class CarManagementTest {
         dataBase = null;
         carListTestCars = null;
         activity = null;
-        management.setCars(new ArrayList<Car>());
     }
 }
