+++ app/src/test/java/se/healthrover/entities/CarManagementImpTest.java
@@ -1,25 +1,46 @@
 package se.healthrover.entities;
 
 
+import com.github.javafaker.Faker;
+
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import se.healthrover.car_service.CarManagement;
 import se.healthrover.car_service.CarManagementImp;
+import se.healthrover.conectivity.HealthRoverWebService;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
 
 public class CarManagementImpTest {
 
 
     private CarManagement carManagementSpy;
+    private HealthRoverWebService healthRoverCar;
+    private CarManagementImp management;
     private Car healthRover;
+    private List<Car> carListTestCars;
+
 
 
     @Before
     public void setUp(){
+        management = new CarManagementImp();
         healthRover = new TestCar();
         CarManagement carManagement = new CarManagementImp();
         carManagementSpy = Mockito.spy(carManagement);
+        carListTestCars = new ArrayList<>();
+        for (int i = 0; i < 10; i++){
+            Car car = new Car("http://" + new Faker().internet().url(), new Faker().name().username());
+            carListTestCars.add(car);
+            management.addCar(car);
+        }
     }
 
 
@@ -36,6 +57,54 @@ public class CarManagementImpTest {
     }
 
 
+    @Test
+    public void getCarByNameTest(){
+        Car car = carListTestCars.get(0);
+        String carName = car.getName();
+        Car result = management.getCarByName(carName);
+        assertEquals(carName, result.getName());
+        assertEquals(car, result);
+    }
+    @Test
+    public void getCarsTest(){
+        List<Car> cars = management.getCars();
+        assertEquals(carListTestCars, cars);
+    }
+
+    @Test
+    public void addCarTest(){
+        Car car = new Car("http://" + new Faker().internet().url(), new Faker().name().username());
+        carListTestCars.add(car);
+        int startSize = management.getCars().size();
+        management.addCar(car);
+        assertEquals(startSize + 1 , management.getCars().size());
+        assertEquals(carListTestCars, management.getCars());
+    }
+
+    @Test
+    public void removeCarTest(){
+        Car car = management.getCars().get(0);
+        carListTestCars.remove(car);
+        int startSize = management.getCars().size();
+        management.removeCar(car);
+        assertEquals(startSize - 1 , management.getCars().size());
+        assertEquals(carListTestCars, management.getCars());
+    }
+    @Test
+    public void setCarListTest(){
+        for (int i = 1; i < management.getCars().size(); i++){
+            management.getCars().remove(i);
+        }
+        assertNotEquals(carListTestCars, management.getCars());
+        management.setCars(carListTestCars);
+        assertEquals(carListTestCars, management.getCars());
+    }
+
+    @After
+    public void tearDown(){
+        carListTestCars = null;
+        management.setCars(new ArrayList<Car>());
+    }
 
 
 }
